requirements.txt
flask
pywebview
pycryptodome